Bảo mật API dùng header token trong các request

+++ Shop_API/Controllers/ImagesController.cs
@@ -1,12 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using Shop_API.AppDbContext;
-using Shop_API.Repository;
+﻿using Microsoft.AspNetCore.Mvc;
 using Shop_API.Repository.IRepository;
 using Shop_Models.Entities;
 
@@ -17,25 +9,51 @@ namespace Shop_API.Controllers
     public class ImagesController : ControllerBase
     {
         private readonly IImageRepository _repository;
+        private readonly IConfiguration _config;
 
-        public ImagesController(IImageRepository repository)
+        public ImagesController(IImageRepository repository, IConfiguration config)
         {
             _repository = repository;
+            _config = config;
         }
 
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Image>>> GetImages()
         {
+
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
             return Ok(await _repository.GetAllImage());
         }
 
         [HttpPut]
-        public async Task<IActionResult> UpdateImage( Image image)
+        public async Task<IActionResult> UpdateImage(Image image)
         {
-            if(await _repository.Update(image))
+
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
+            if (await _repository.Update(image))
             {
                 return Ok("Sửa thành công");
-               
+
             }
             return BadRequest("Sửa thất bại");
         }
@@ -45,8 +63,20 @@ namespace Shop_API.Controllers
         [HttpPost]
         public async Task<ActionResult<Image>> CreateImage(Image image)
         {
-          image.Id= Guid.NewGuid();
-          if(await _repository.Create(image))
+
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
+            image.Id = Guid.NewGuid();
+            if (await _repository.Create(image))
             {
                 return Ok("Thêm thành công");
             }
@@ -57,13 +87,26 @@ namespace Shop_API.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteImage(Guid id)
         {
-           if(await _repository.Delete(id)) {
+
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
+            if (await _repository.Delete(id))
+            {
 
                 return Ok("Xóa thành công");
-           }
+            }
             return BadRequest("Xóa thất bại");
         }
 
-      
+
     }
 }
