+++ Shop_API/Controllers/ProductTypeController.cs
@@ -9,14 +9,44 @@ namespace Shop_API.Controllers
     public class ProductTypeController : ControllerBase
     {
         private readonly IProductTypeRepository _repository;
-        public ProductTypeController(IProductTypeRepository repository)
+        private readonly IConfiguration _config;
+        public ProductTypeController(IProductTypeRepository repository, IConfiguration config)
         {
             _repository = repository;
+            _config = config;
         }
+        [HttpGet]
+        public async Task<IActionResult> GetAllProductType()
+        {
 
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
+            return Ok(await _repository.GetAll());
+        }
         [HttpPost]
         public async Task<IActionResult> CreateProductType(ProductType obj)
         {
+
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
             obj.Id = Guid.NewGuid();
             if (await _repository.Create(obj))
             {
@@ -29,6 +59,17 @@ namespace Shop_API.Controllers
         public async Task<IActionResult> UpdateProductType(ProductType obj)
         {
 
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
             if (await _repository.Update(obj))
             {
                 return Ok("Chỉnh sửa thành công");
@@ -40,6 +81,17 @@ namespace Shop_API.Controllers
         public async Task<IActionResult> DeleteProductType(Guid id)
         {
 
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
             if (await _repository.Delete(id))
             {
                 return Ok("Xóa thành công");
@@ -47,10 +99,6 @@ namespace Shop_API.Controllers
             return BadRequest("Xóa thất bại");
         }
 
-        [HttpGet]
-        public async Task<IActionResult> GetAllProductType()
-        {
-            return Ok(await _repository.GetAll());
-        }
+
     }
 }
