+++ Shop_API/Controllers/SerialDaBanController.cs
@@ -9,18 +9,44 @@ namespace Shop_API.Controllers
     public class SerialDaBanController : ControllerBase
     {
         private readonly ISerialDaBanRepository _repository;
-        public SerialDaBanController(ISerialDaBanRepository repository)
+        private readonly IConfiguration _config;
+        public SerialDaBanController(ISerialDaBanRepository repository, IConfiguration config)
         {
             _repository = repository;
+            _config = config;
         }
         [HttpGet]
         public async Task<IActionResult> GetAllSerialDaBan()
         {
+
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
             return Ok(await _repository.GetAll());
         }
         [HttpPost]
         public async Task<IActionResult> CreateSerialDaBan(SerialDaBan obj)
         {
+
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
             obj.Id = Guid.NewGuid();
             if (await _repository.Create(obj))
             {
@@ -32,6 +58,18 @@ namespace Shop_API.Controllers
         [HttpPut]
         public async Task<IActionResult> UpdateSerialDaBan(SerialDaBan obj)
         {
+
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
             if (await _repository.Update(obj))
             {
                 return Ok("Chỉnh sửa thành công");
@@ -41,6 +79,18 @@ namespace Shop_API.Controllers
         [HttpDelete("id")]
         public async Task<IActionResult> DeleteSerialDaBan(Guid id)
         {
+
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                return Unauthorized();
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                return Unauthorized();
+            }
             if (await _repository.Delete(id))
             {
                 return Ok("Xóa thành công");
