+++ Shop_API/Controllers/ViDiemController.cs
@@ -11,14 +11,36 @@ namespace Shop_API.Controllers
     {
         private readonly IViDiemRepository _viDiemRepository;
         private ReponseDto _reponseDto;
-        public ViDiemController(IViDiemRepository viDiemRepository)
+        private readonly IConfiguration _config;
+        public ViDiemController(IViDiemRepository viDiemRepository, IConfiguration config)
         {
             _viDiemRepository = viDiemRepository;
             _reponseDto = new ReponseDto();
+            _config = config;
         }
         [HttpGet]
         public async Task<ReponseDto> GetAllViDiem()
         {
+
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                _reponseDto.Result = null;
+                _reponseDto.IsSuccess = false;
+                _reponseDto.Code = 401;
+                _reponseDto.Message = "Không có quyền";
+                return _reponseDto;
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                _reponseDto.Result = null;
+                _reponseDto.IsSuccess = false;
+                _reponseDto.Code = 401;
+                _reponseDto.Message = "Không có quyền";
+                return _reponseDto;
+            }
             var list = await _viDiemRepository.GetAllViDiems();
             if (list != null)
             {
@@ -38,6 +60,25 @@ namespace Shop_API.Controllers
         [HttpPost]
         public async Task<ReponseDto> CreateViDiem(ViDiem obj)
         {
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                _reponseDto.Result = null;
+                _reponseDto.IsSuccess = false;
+                _reponseDto.Code = 401;
+                _reponseDto.Message = "Không có quyền";
+                return _reponseDto;
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                _reponseDto.Result = null;
+                _reponseDto.IsSuccess = false;
+                _reponseDto.Code = 401;
+                _reponseDto.Message = "Không có quyền";
+                return _reponseDto;
+            }
             if (await _viDiemRepository.Create(obj))
             {
                 _reponseDto.Result = obj;
@@ -56,6 +97,25 @@ namespace Shop_API.Controllers
         [HttpPut]
         public async Task<ReponseDto> UpdateViDiem(ViDiem obj)
         {
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                _reponseDto.Result = null;
+                _reponseDto.IsSuccess = false;
+                _reponseDto.Code = 401;
+                _reponseDto.Message = "Không có quyền";
+                return _reponseDto;
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                _reponseDto.Result = null;
+                _reponseDto.IsSuccess = false;
+                _reponseDto.Code = 401;
+                _reponseDto.Message = "Không có quyền";
+                return _reponseDto;
+            }
             if (await _viDiemRepository.Update(obj))
             {
                 _reponseDto.Result = obj;
@@ -75,6 +135,25 @@ namespace Shop_API.Controllers
         [HttpDelete]
         public async Task<ReponseDto> DeleteViDiem(Guid id)
         {
+            string apiKey = _config.GetSection("ApiKey").Value;
+            if (apiKey == null)
+            {
+                _reponseDto.Result = null;
+                _reponseDto.IsSuccess = false;
+                _reponseDto.Code = 401;
+                _reponseDto.Message = "Không có quyền";
+                return _reponseDto;
+            }
+
+            var keyDomain = Request.Headers["Key-Domain"].FirstOrDefault();
+            if (keyDomain != apiKey.ToLower())
+            {
+                _reponseDto.Result = null;
+                _reponseDto.IsSuccess = false;
+                _reponseDto.Code = 401;
+                _reponseDto.Message = "Không có quyền";
+                return _reponseDto;
+            }
             if (await _viDiemRepository.Delete(id))
             {
                 _reponseDto.Result = null;
